PoC implementation of special metadata for issues
Add additional metadata for the issues to enable some (optional) automatic behavior during the submission process

A new (optional) step is added to allow user to select the issue to submit to if this is enabled by the journal on an issue by issue basis.

If an issue is marked as "Collection" additional fields can be used to "guide" the submission process.

Each case below is optional and they can be used together to implement different submission models

- date based selection: issues are available for selection only if the submission is within the submission timeframe set in the issue (fields: date_open / date_close)
- invitation based selection: issues are available for selection only if the user is among a list on invited users
- sections restriction: sections available in the submit_info step are filtered according to the (optional) issue selected
- automatic editor selection: articles can be assigned automatically to a set of editors assigned to the issue (not implemented yet)

Add additional fields to issue to handle presentation and internal data